Fix framebuffer sizing on linux, need to make the API easier to work with for different DPI scales.

+++ core/include/vara/core/math/vec4.h
@@ -4,7 +4,7 @@
 
 // Constructors
 
-static Vector4 vec4(f32 x, f32 y, f32 z, i32 w) {
+static Vector4 vec4(f32 x, f32 y, f32 z, f32 w) {
     return (Vector4){
         x,
         y,
